added basic test for parser

+++ lib/folder_file_manager/folder_file_manager.cpp
@@ -43,14 +43,22 @@ void FolderFileManager::_init()
 FolderFileManager::~FolderFileManager(){}
 
 
-bool FolderFileManager::file_exist(const string& path)
+bool FolderFileManager::file_exist(string path)
 {
+    int idx = first_not_space(path, 0);
+    if(idx == -1) return false;
+    path = path.substr(idx);
+
     if(this->_file_set.count(this->_base_dir + path) > 0) return true;
     return false;
 }
 
-bool FolderFileManager::folder_exist(const string& path)
+bool FolderFileManager::folder_exist(string path)
 {
+    int idx = first_not_space(path, 0);
+    if(idx == -1) return false;
+    path = path.substr(idx);
+
     if(this->_folder_set.count(this->_base_dir + path + "/") > 0) return true;
     return false;
 }
@@ -64,7 +72,7 @@ bool FolderFileManager::add_folder(string path)
 {
     int idx = first_not_space(path, 0);
     if(idx == -1) return false;
-    path = path.substr(idx, path.length());
+    path = path.substr(idx);
 
     ofstream outs;
     string file_path = this->_base_dir + path + "/";
@@ -83,7 +91,7 @@ bool FolderFileManager::add_file(string path)
 {
     int idx = first_not_space(path, 0);
     if(idx == -1) return false;
-    path = path.substr(idx, path.length());
+    path = path.substr(idx);
     
     ofstream outs;
     string file_path = this->_base_dir + path;
@@ -125,8 +133,12 @@ void FolderFileManager::_add_delete_file_folder(const string& file_path, bool co
 
 
 
-bool FolderFileManager::delete_folder(const string& path)
+bool FolderFileManager::delete_folder(string path)
 {
+    int idx = first_not_space(path, 0);
+    if(idx == -1) return false;
+    path = path.substr(idx);
+
     string file_path = this->_base_dir + path + "/";
     if(!this->_folder_set.count(file_path)) return false;
     if(rmdir(file_path.c_str()) == -1) return false;
@@ -140,8 +152,12 @@ bool FolderFileManager::delete_folder(const string& path)
 }
 
 
-bool FolderFileManager::delete_file(const string& path)
+bool FolderFileManager::delete_file(string path)
 {
+    int idx = first_not_space(path, 0);
+    if(idx == -1) return false;
+    path = path.substr(idx);
+    
     string file_path = this->_base_dir + path;
     if(!this->_file_set.count(file_path)) return false;
     if(remove(file_path.c_str()) == -1) return false;
